+++ benchmark/src/main/java/io/github/pr0methean/betterrandom/benchmark/AesCounterRandom128Benchmark.java
@@ -1,5 +1,7 @@
 package io.github.pr0methean.betterrandom.benchmark;
 
+import static io.github.pr0methean.betterrandom.seed.SecureRandomSeedGenerator.SECURE_RANDOM_SEED_GENERATOR;
+
 import io.github.pr0methean.betterrandom.prng.AesCounterRandom;
 import io.github.pr0methean.betterrandom.seed.SeedException;
 import java.util.Random;
@@ -7,6 +9,6 @@ import java.util.Random;
 public class AesCounterRandom128Benchmark extends AbstractRandomBenchmarkWithReseeding {
 
   @Override protected Random createPrng() throws SeedException {
-    return new AesCounterRandom(16);
+    return new AesCounterRandom(SECURE_RANDOM_SEED_GENERATOR.generateSeed(16));
   }
 }