+++ benchmark/src/main/java/io/github/pr0methean/betterrandom/benchmark/CellularAutomatonRandomBenchmark.java
@@ -1,5 +1,7 @@
 package io.github.pr0methean.betterrandom.benchmark;
 
+import static io.github.pr0methean.betterrandom.seed.SecureRandomSeedGenerator.SECURE_RANDOM_SEED_GENERATOR;
+
 import io.github.pr0methean.betterrandom.prng.CellularAutomatonRandom;
 import io.github.pr0methean.betterrandom.seed.SeedException;
 import java.util.Random;
@@ -7,6 +9,6 @@ import java.util.Random;
 public class CellularAutomatonRandomBenchmark extends AbstractRandomBenchmarkWithReseeding {
 
   @Override protected Random createPrng() throws SeedException {
-    return new CellularAutomatonRandom();
+    return new CellularAutomatonRandom(SECURE_RANDOM_SEED_GENERATOR);
   }
 }