+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # 2.5.0
 * Adds `Pcg128Random`, a variant of `Pcg64Random` that uses a 128-bit seed and state.
 
+# 2.4.4
+* `RandomSeederThread` no longer overrides `Thread.interrupt()`.
+* `getSeedGenerator` now behaves as specified when called on a `ReseedingSplittableRandomAdapter` or
+  `ReseedingThreadLocalRandomWrapper`.
+
 # 2.4.3
 * Reduces unit-test flakiness further on systems where `DefaultSeedGenerator` is slow.